+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* The `showFiltersInModal` property in the `flexberry-objectlistview` component for displayng filters in a modal window.
+* The `errorMessageFilterActive` and `errorMessageFilters` properties for manage the filtering of errors in the `log` service.
 
 ## [2.4.0-beta.2] - 2020-03-26
 ### Fixed